refactor(runtime): move LoadTeamRequest and CreateSessionRequest to pkg/runtime
The two structs lived in cmd/root because LocalBackend was the only
consumer. They are about to gain JSON tags and a wire-protocol round
trip; the natural home for that is pkg/runtime, alongside the Runtime
interface they parametrize.

cmd/root keeps the factory methods (loadTeamRequest /
createSessionRequest) — those still read from the CLI flag struct
and stay near it. The struct types themselves move; consumers now
import them as runtime.LoadTeamRequest / runtime.CreateSessionRequest.

No behavioural change; sets up the next step's JSON tags + round-trip
tests.
